changed beeHappy effect to double

+++ src/main/kotlin/de/unisaarland/cs/se/selab/incidents/BeeHappy.kt
@@ -11,7 +11,7 @@ class BeeHappy(
     id: Int,
     tick: Int,
     val affectedTiles: Set<Tile>,
-    val effect: Int,
+    val effect: Double,
     private val yearTick: Int
 ) : Incident(id, tick) {
